requirements.txt
flask>=3.0.0
flask-cors>=5.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-anthropic>=0.2.0
langchain-groq>=0.2.0
langgraph>=0.2.0
python-docx>=1.1.0
reportlab>=4.1.0
structlog>=24.1.0
